searchbar modified
