+++ views/layouts/main.handlebars
@@ -16,7 +16,9 @@
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@3.3.7/dist/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
 <link rel="stylesheet" href="/public/css/styles.css">
 <link rel="shortcut icon" href="#">
+<script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
 <script src="/public/js/client_side_validation.js"></script>
+<script src="/public/js/propertySearchAjax.js"></script>
 </head>
 <body>
 <header>
@@ -25,10 +27,20 @@
 <h1><a href="/">Apartment Hunt Helper</a></h1>
 </div>
 {{#if userData}}
-<form id="searchForm" action="/searchProperty" method="GET">
-<label for="searchProperty"></label>
-<input class="search-input" type="text" id="searchProperty" name="name" placeholder="Search for Properties">
-<button class="search-button" type="submit" value="Find Property">Search</button>
+<form id="searchForm" method="GET">
+  <label for="searchType">Search Type:</label>
+  <select id="searchType" name="type">
+    <option value="">Select filter</option>
+      <option value="name">Name</option>
+    <option value="address">Address</option>
+    <option value="city">City</option>
+    <option value="state">State</option>
+    <option value="zip">Zip Code</option>
+  </select>
+  <label for="searchQuery">Search Query:</label>
+  <input class="search-input" type="text" id="searchQuery" name="query" placeholder="search here...">
+  <button class="search-button" type="submit">Search</button>
+<div id='errorDiv' hidden></div>
 </form>
 {{/if}}
 <div class="user-container">
@@ -42,6 +54,9 @@
 </header>
 
 <main>
+    <ul id="searchResults" hidden></ul>
+    <div id="propertyDetails" hidden></div>
+    <a id="rootLink" href="/" hidden>New Search</a>
 {{{body}}}
 </main>
